// WebApi/ClientApp/src/models/responses/category/IDonVi.ts
export interface IDonVi {
    id: number;
    ma_dv: string;
    ten_dv: string;
    dia_chi: string;
    mst: string;
    stk: string;
    dien_thoai: string;
    ghi_chu: string;
    donvi_chuquan: string;
    co_quan_thu_id_chuquan: number;
    ngan_hang: string;
    fax: string;
    website: string;
    email: string;
    ma_dang_ky_cqt?: string;
    ngay_hoa_don_max?: string;
    to_khai_success_id?: number;
    serials:string;
}